+++ src/lib/components/CurrentPlayingSongs.svelte
@@ -1,5 +1,7 @@
 <script lang="ts">
-  import { currentSongs } from '$lib/store'
+  import { db } from '$lib/db'
+  import { currentPlayingSongIds } from '$lib/store'
+  import type { Song } from '$lib/types'
   import Backdrop from './Backdrop.svelte'
   import { fullscreen } from './lyrics/store'
   import Songs from './Songs.svelte'
@@ -8,19 +10,37 @@
 
   const handleRemoveFromList = (e: CustomEvent<number>) => {
     const songIdToRemove = e.detail
-    const idx = $currentSongs.findIndex(song => song.id === songIdToRemove)
+    const idx = $currentPlayingSongIds.findIndex(id => id === songIdToRemove)
     if (idx !== -1) {
-      $currentSongs = [
-        ...$currentSongs.slice(0, idx),
-        ...$currentSongs.slice(idx + 1),
+      $currentPlayingSongIds = [
+        ...$currentPlayingSongIds.slice(0, idx),
+        ...$currentPlayingSongIds.slice(idx + 1),
       ]
     }
+  }
+
+  let paginatedSongs: Song[] = []
+
+  let offset = 0
+  let limit = 10
+
+  const doPaginateSongs = async () => {
+    paginatedSongs = await db.songs.offset(offset).limit(limit).toArray()
+  }
+
+  $: {
+    offset
+    limit
+    doPaginateSongs()
   }
 </script>
 
 <div class="current-songs" class:show="{show}" class:fullscreen="{$fullscreen}">
   <Songs
-    songs="{$currentSongs}"
+    bind:limit="{limit}"
+    bind:offset="{offset}"
+    songs="{paginatedSongs}"
+    total="{$currentPlayingSongIds.length}"
     transparentBg="{$fullscreen}"
     resetCurrentSongsOnClick="{false}"
     showActionsOnEmpty="{false}"