+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Resize observe use debounce
 * Title bar and player control bottom bar auto hide in full screen mode
 * Volume control dark mode style
+* Optimize all queries with virtual scroll + pagination
 
 ### 🐛 Fixes